requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
Arpeggio==2.0.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
bitarray==3.8.0
bitmath==1.3.3.1
bitstring==4.3.1
blessed==1.17.12
bpylist2==4.1.1
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cgmetadata==0.2.0
charset-normalizer==3.4.4
chromadb==1.3.4
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.3
distro==1.9.0
durationpy==0.10
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
mac-alias==2.2.2
makelive==0.6.2
Mako==1.2.4
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
mcp==1.21.1
mdurl==0.1.2
mmh3==5.2.0
more-itertools==8.14.0
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
objexplore==1.6.3
ollama==0.6.1
onnxruntime==1.23.2
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
osxmetadata==1.4.0
osxphotos==0.74.0
overrides==7.7.0
packaging==25.0
parso==0.8.5
pathvalidate==3.3.1
photoscript==0.5.2
pillow==12.0.0
posthog==5.4.0
prompt_toolkit==3.0.52
protobuf==6.33.1
ptpython==3.0.31
py-applescript==1.0.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyobjc-core==12.1
pyobjc-framework-AppleScriptKit==12.1
pyobjc-framework-AppleScriptObjC==12.1
pyobjc-framework-AVFoundation==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Contacts==12.1
pyobjc-framework-CoreAudio==12.1
pyobjc-framework-CoreLocation==12.1
pyobjc-framework-CoreMedia==12.1
pyobjc-framework-CoreML==12.1
pyobjc-framework-CoreServices==12.1
pyobjc-framework-FSEvents==12.1
pyobjc-framework-Metal==12.1
pyobjc-framework-Photos==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-ScriptingBridge==12.1
pyobjc-framework-UniformTypeIdentifiers==12.1
pyobjc-framework-Vision==12.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytimeparse2==1.7.1
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==13.9.4
rich-theme-manager==0.11.0
rpds-py==0.29.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.3
starlette==0.50.0
strpdatetime==0.4.1
sympy==1.14.0
tenacity==8.5.0
textX==4.2.3
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.2.2
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
utitools==0.4.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
whenever==0.8.10
wrapt==1.17.3
wurlitzer==3.1.1
xattr==1.3.0
xdg-base-dirs==6.0.2
zipp==3.23.0
